+++ changelog.md
@@ -2,3 +2,4 @@ Changelog next version:
 
 * Catchup subscriptions (blocking)
 * EveryN for stream persistence (both blocking and reactive)
+* Added "count" to EventStoreQueries (both blocking and reactive)